fix pre-commit

+++ torchao/quantization/quantize_/workflows/int8/int8_tensor.py
@@ -327,27 +327,31 @@ def _(func, types, args, kwargs):
                 w_scales.reshape(-1, 1) if w_scales.numel() > 1 else w_scales,
             )
         elif kernel_choice == KernelPreference.TORCH:
-          # Cast FP16 scale to float to avoid overflow in int_scaled_matmul
-          intermediate_dtype = (
-              torch.float if x_scales.dtype == torch.half else x_scales.dtype
-          )
-          y_dot_scaled = int_scaled_matmul(
-              tmp, w_vals_int8_t, x_scales.reshape(-1, 1).to(intermediate_dtype)
-          ).to(output_dtype)
+            # Cast FP16 scale to float to avoid overflow in int_scaled_matmul
+            intermediate_dtype = (
+                torch.float if x_scales.dtype == torch.half else x_scales.dtype
+            )
+            y_dot_scaled = int_scaled_matmul(
+                tmp, w_vals_int8_t, x_scales.reshape(-1, 1).to(intermediate_dtype)
+            ).to(output_dtype)
 
-          # Asymmetric activation zero_point correction:
-          # Y = (X_int @ W_int^T) * s_x * s_w - zp_x * s_x * row_sum(W_int)^T * s_w
-          # The first term is y_dot_scaled * w_scales. The second is the correction.
-          if activation_tensor.zero_point is not None:
-              w_row_sums = weight_tensor.qdata.sum(dim=-1)  # (N,)
-              zp_x = activation_tensor.zero_point.reshape(-1, 1).to(intermediate_dtype)
-              x_scales_flat = x_scales.reshape(-1, 1).to(intermediate_dtype)
-              zp_correction = (zp_x * x_scales_flat) * w_row_sums.to(intermediate_dtype)
-              y_dot_scaled = y_dot_scaled - zp_correction.to(output_dtype)
+            # Asymmetric activation zero_point correction:
+            # Y = (X_int @ W_int^T) * s_x * s_w - zp_x * s_x * row_sum(W_int)^T * s_w
+            # The first term is y_dot_scaled * w_scales. The second is the correction.
+            if activation_tensor.zero_point is not None:
+                w_row_sums = weight_tensor.qdata.sum(dim=-1)  # (N,)
+                zp_x = activation_tensor.zero_point.reshape(-1, 1).to(
+                    intermediate_dtype
+                )
+                x_scales_flat = x_scales.reshape(-1, 1).to(intermediate_dtype)
+                zp_correction = (zp_x * x_scales_flat) * w_row_sums.to(
+                    intermediate_dtype
+                )
+                y_dot_scaled = y_dot_scaled - zp_correction.to(output_dtype)
 
-          y = (y_dot_scaled * w_scales.flatten()).reshape(
-              *x_vals_int8.shape[:-1], y_dot_scaled.shape[-1]
-          )
+            y = (y_dot_scaled * w_scales.flatten()).reshape(
+                *x_vals_int8.shape[:-1], y_dot_scaled.shape[-1]
+            )
 
     else:
         # FP × INT8 (weight-only)
@@ -506,7 +510,6 @@ def _(func, types, args, kwargs):
         act_pre_scale=old_int8_tensor.act_pre_scale,
         act_quant_kwargs=old_int8_tensor.act_quant_kwargs,
         kernel_preference=old_int8_tensor.kernel_preference,
-
     )
     return return_and_correct_aliasing(func, args, kwargs, new_int8_tensor)
